Some logic moved to css plus some refactoring
