+++ kafka-ui-react-app/src/components/Topics/Topic/Details/Details.styled.ts
@@ -6,3 +6,13 @@ export const DropdownExtraMessage = styled.div`
   width: 100%;
   margin-top: 10px;
 `;
+
+export const ReplicaCell = styled.span.attrs({ 'aria-label': 'replica-info' })<{
+  leader: boolean | undefined;
+}>`
+  ${this} ~ ${this}::before {
+    color: black;
+    content: ', ';
+  }
+  color: ${(props) => (props.leader ? 'orange' : null)};
+`;
